+++ docker/startup/start_git.sh
@@ -1,0 +1,8 @@
+#!/usr/bin/env bash
+
+cd /opt/ace
+source /venv/bin/activate
+source load_environment
+source docker/startup/start.sh
+
+ace -L etc/logging_configs/service_git.ini service start git